+++ client/components/experiment/experiment-props.ts
