+++ client/components/experiment/index.tsx
@@ -10,11 +10,11 @@ import { AppState } from 'types';
  */
 import { getVariationForUser, isLoading } from 'state/experiments/selectors';
 import QueryExperiments from 'components/data/query-experiments';
-import { ExperimentProps } from './experimentProps';
+import { ExperimentProps } from './experiment-props';
 
 export { default as Variation } from './variation';
-export { default as DefaultVariation } from './defaultVariation';
-export { default as LoadingVariations } from './loadingVariations';
+export { default as DefaultVariation } from './default-variation';
+export { default as LoadingVariations } from './loading-variations';
 
 /**
  * The experiment component to display the experiment variations
