+++ client/components/experiment/variation-props.ts
